+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## unreleased
+### Changed
+* Adjusted material list to wrap long material labels in Matl Editor and Material Preset Editor.
 
 ## 0.10.3 - 2024-05-15
 ### Fixed